+++ src/tests.rs
@@ -488,7 +488,7 @@ where
         1 => l.insert_generic::<String>(&val).unwrap(),
         2 => l
           .insert_with(ValueBuilder::new(
-            val.len() as u32,
+            val.len(),
             |buf: &mut VacantBuffer<'_>| buf.put_slice(val.as_bytes()),
           ))
           .unwrap(),
@@ -496,7 +496,7 @@ where
         4 => l.insert_generic_tombstone::<String>(&val).unwrap(),
         5 => l
           .insert_tombstone_with(ValueBuilder::new(
-            val.len() as u32,
+            val.len(),
             |buf: &mut VacantBuffer<'_>| buf.put_slice(val.as_bytes()),
           ))
           .unwrap(),
@@ -549,7 +549,7 @@ where
         1 => l.insert_generic::<String>(&val).unwrap(),
         2 => l
           .insert_with(ValueBuilder::new(
-            val.len() as u32,
+            val.len(),
             |buf: &mut VacantBuffer<'_>| buf.put_slice(val.as_bytes()),
           ))
           .unwrap(),
@@ -557,7 +557,7 @@ where
         4 => l.insert_generic_tombstone::<String>(&val).unwrap(),
         5 => l
           .insert_tombstone_with(ValueBuilder::new(
-            val.len() as u32,
+            val.len(),
             |buf: &mut VacantBuffer<'_>| buf.put_slice(val.as_bytes()),
           ))
           .unwrap(),
